+++ src/components/header/extendedHeader/ExtendedHeader/ExtendedHeader.stories.tsx
@@ -1,7 +1,7 @@
 import React, { useState } from 'react'
 import { ExtendedHeader } from './ExtendedHeader'
 
-import { Title } from '../../../Title/Title'
+import { Title } from '../../Title/Title'
 import { ExtendedNav } from '../ExtendedNav/ExtendedNav'
 import { SearchInput } from '../../../SearchInput/SearchInput'
 import { DropDownLink } from '../../DropDownMenu/DropDownMenu'
@@ -85,7 +85,7 @@ export const basicExtendedHeader = (): React.ReactElement => {
       <div className={`usa-overlay ${expanded ? 'is-visible' : ''}`}></div>
       <ExtendedHeader>
         <div className="usa-navbar">
-          <Title title="Project Title" link="#testlink"></Title>
+          <Title>Project Title</Title>
           <NavButton onClick={onClick} className="usa-menu-btn">
             Menu
           </NavButton>
@@ -110,7 +110,7 @@ export const extendedHeaderWithMegaMenu = (): React.ReactElement => {
       <div className={`usa-overlay ${expanded ? 'is-visible' : ''}`}></div>
       <ExtendedHeader>
         <div className="usa-navbar">
-          <Title title="Project Title" link="#testlink"></Title>
+          <Title>Project Title</Title>
           <NavButton onClick={onClick} className="usa-menu-btn">
             Menu
           </NavButton>
